+++ app/Http/Controllers/API/ExpenditureController.php
@@ -68,6 +68,9 @@ class ExpenditureController extends Controller
     public function destroy($id)
     {
         $department = expenditure::findOrFail($id);
+        if ($department->debit->count() > 0){
+            return ['error'=>'This income source all ready used, you can\'t delete'];
+        }
         $department->delete();
         return ['message'=>'User deleted'];
     }
